refactor(aggregator-test): extract leader follower test http server to a module
In integration tests
